JS-8974: skip .desktop file write on Linux if it already exists
Prevent overwriting user customizations to anytype.desktop by only
writing the file on first launch (when it doesn't exist yet).

+++ electron/js/util.js
@@ -337,20 +337,22 @@ class Util {
 		].join('\n');
 
 		try {
-			fs.mkdirSync(applicationsDir, { recursive: true });
-			fs.writeFileSync(desktopFilePath, content, 'utf-8');
+			if (!fs.existsSync(desktopFilePath)) {
+				fs.mkdirSync(applicationsDir, { recursive: true });
+				fs.writeFileSync(desktopFilePath, content, 'utf-8');
 
-			execFile('update-desktop-database', [ applicationsDir ], (err) => {
-				if (err) {
-					this.log('info', `update-desktop-database failed: ${err.message}`);
-				};
-			});
+				execFile('update-desktop-database', [ applicationsDir ], (err) => {
+					if (err) {
+						this.log('info', `update-desktop-database failed: ${err.message}`);
+					};
+				});
 
-			execFile('xdg-mime', [ 'default', 'anytype.desktop', 'x-scheme-handler/anytype' ], (err) => {
-				if (err) {
-					this.log('info', `xdg-mime default failed: ${err.message}`);
-				};
-			});
+				execFile('xdg-mime', [ 'default', 'anytype.desktop', 'x-scheme-handler/anytype' ], (err) => {
+					if (err) {
+						this.log('info', `xdg-mime default failed: ${err.message}`);
+					};
+				});
+			};
 		} catch (e) {
 			this.log('info', `registerLinuxProtocolHandler failed: ${e.message}`);
 		};
